requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
assemblyai==0.17.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
cohere==4.23
colorama==0.4.6
comm==0.1.4
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
fastavro==1.8.2
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
langchain==0.0.283
langsmith==0.0.33
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.0
numexpr==2.8.5
numpy==1.25.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
PyAudio==0.2.13
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
SQLAlchemy==2.0.20
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.3
tokenizers==0.13.3
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@2af87d018eff577c421b02e00bf34ab6aeeb8c42
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
